Removing Lotus app support for CLI Update Lotus::Model::Migrator setup Update documentation
