+++ lib/lotus/commands/db/migrator.rb
@@ -1,4 +1,5 @@
-require 'lotus/config/adapter'
+require 'lotus/model/migrator'
+require 'lotus/model/migration'
 
 module Lotus
   module Commands
@@ -12,10 +13,9 @@ module Lotus
       #   `bundle exec lotus db migrate`
       #   `bundle exec lotus db rollback`
       #
-      # Allows user to specify lotus app and number of steps in case of rollback
+      # Allows user to specify number of steps when rollback
       #
-      #   `bundle exec lotus db migrate web`
-      #   `bundle exec lotus db rollback web --step=2`
+      #   `bundle exec lotus db rollback -step=2`
       #
       # This feature uses Lotus::Model::Migratior to execute migrations 
       #
@@ -25,57 +25,41 @@ module Lotus
 
         # Default number of STEPS to roll back 
         DEFAULT_STEP = 1 
-        APPS_DIRECOTRY = "apps".freeze
-        MIGRATION_DIRECTORY = "db/migrations".freeze
 
-        def initialize(name=nil, environment)
-          @name         = name
+        def initialize(environment)
           @environment  = environment
           @options      = @environment.to_options
-          @adapter      = Lotus::Config::Adapter.new(@name)
           load_environment
-          @migrator     = Lotus::Model::Migrator.new(@adapter.config, logger: Lotus::Logger.new)
+          @migrator     = Lotus::Model::Migrator.new
         end
 
-        # Apply all migrations up located on specified directory
-        # the directories could be located on root/db/migrations
-        # or root/apps/#{app}/db/migrations
+        # Apply all migrations by calling Lotus::Model::Migrator
+        #
+        # @since 0.3.0
+        # @api private
+        # @see Lotus::Model::Migrator#migrate
         def migrate
-          @migrator.migrate(directory: migration_directory)
+          @migrator.migrate
         end
 
-        # Apply migrations down located on specified directory
-        # the directories could be  root/db/migrations
-        # or root/apps/#{app}/db/migrations
-        # A number of steps could be specified by `--step` option
+        # Apply all migrations by calling Lotus::Model::Migrator
+        #
+        # @since 0.3.0
+        # @api private       
+        # @see Lotus::Model::Migrator#rollback
         def rollback
-          @migrator.rollback(directory: migration_directory, step: step)
+          @migrator.rollback(step: step)
         end
 
         private 
-        attr_reader :name
-
-        # It retuns the default migration directory or lotus  
-        # app migration 
-        def migration_directory
-          if name 
-            pwd.join(APPS_DIRECOTRY, name, MIGRATION_DIRECTORY)
-          else
-            pwd.join(MIGRATION_DIRECTORY)
-          end
-        end
-
-        def adapter_config
-        end
-        
+        # Get number of steps
+        #
+        # @since 0.3.0
+        # @api private
         def step
           @options.fetch(:step, DEFAULT_STEP) 
         end
-         
-        def pwd
-          @pwd ||= Pathname.new(Dir.pwd)
-        end
-        
+       
         def load_environment
           require @environment.env_config
         end
