+++ test/fixtures/sql_adapter.rb
@@ -2,18 +2,21 @@ require 'lotus'
 require 'lotus/model'
 require 'lotus/model/adapters/sql_adapter'
 
-db = Pathname.new(__dir__).join('../tmp/db')
-db.dirname.mkpath        # create directory if not exist
+db_path = Pathname.new(__dir__).join('../tmp/db')
+db_path.dirname.mkpath        # create directory if not exist
 
-sql = db.join('sql.db')
+sql = db_path.join('sql.db')
 sql.delete if sql.exist? # delete file if exist
 
 if Lotus::Utils.jruby?
   require 'jdbc/sqlite3'
   Jdbc::SQLite3.load_driver
-  SQLITE_CONNECTION_STRING = "jdbc:sqlite:#{ sql }"
+  SQLITE_CONNECTION_STRING_PATH = "jdbc:sqlite:#{db_path}/"
 else
   require 'sqlite3'
-  SQLITE_CONNECTION_STRING = "sqlite://#{ sql }"
-end 
+  SQLITE_CONNECTION_STRING_PATH = "sqlite://#{db_path}/"
+end
+
+SQLITE_CONNECTION_STRING = "#{SQLITE_CONNECTION_STRING_PATH}/sql.db"
+
